Split TempFileSpace to TempFileSpaceImpl
